// src/app/models/notificacionDTO.ts
export class NotificacionDTO{
  notificacionID?: number;
  tipoDocumento?: string;
  numeroDocumento?: string;
  descripcion?: string;
  codigoNotificacion?: string;
  estadoNotificacion?: number;
  fechaModificacion?: string;
  fechaNotificacion?: string;
  identificador?: string;
  tituloNotificacion?: string;
  usuarioNombre?: string;
  corregido?: boolean;
  checked?: boolean;
  mostrar?: boolean;
}