// src/app/models/usuario.ts
export class Usuario {
  usuarioID?: string;
  apellidos?: string;
  celular?: string;
  email?: string;
  nombres?: string;
  numDoc?: string;
  password?: string;
  tipoDoc?: string;
  captcha?: any;
  tipoDocumento?: string;
  numeroDocumento?: string;
  departamento?: string;
  ciudad?: string;
  direccion?: string;
  nombrePerfil?: string;
  screen?: string;
  updateRedis?: boolean; 
}